프로토콜 위치 수정, 데이터 호출관련 코드 manager로 이동
